base, site and uk_regional _datamodule related updates
